Switch theme to everforest
Lots of good, meaningful updates to my vimrc file.

After months of testing the everforest theme is now my default,
replacing gruvbox but retaining all its settings for future emergency
use.

The Git log library GV is replaced with vim-flog and a handy new mapping
to quickly add the date and time to a file is added.